lens: move v2 utilities to v2_util.go
