+++ v2.go
@@ -2,27 +2,20 @@ package lens
 
 import (
 	"context"
-	"errors"
 	"fmt"
-	"net/http"
 	"strings"
-	"time"
-
-	"google.golang.org/grpc/codes"
-
-	"google.golang.org/grpc/status"
 
 	"go.uber.org/zap"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+
+	"github.com/RTradeLtd/grpc/lensv2"
+	"github.com/RTradeLtd/rtfs"
 
 	"github.com/RTradeLtd/Lens/analyzer/images"
 	"github.com/RTradeLtd/Lens/analyzer/ocr"
 	"github.com/RTradeLtd/Lens/engine"
 	"github.com/RTradeLtd/Lens/xtractor/planetary"
-	"github.com/RTradeLtd/grpc/lensv2"
-	"github.com/RTradeLtd/rtfs"
-
-	"github.com/RTradeLtd/Lens/logs"
-	"github.com/RTradeLtd/Lens/models"
 )
 
 // V2 is the new Lens API, and implements the LensV2 gRPC interface directly.
@@ -113,7 +106,7 @@ func (v *V2) Index(ctx context.Context, req *lensv2.IndexReq) (*lensv2.IndexResp
 
 	var hash = req.GetHash()
 	var reindex = req.GetOptions().GetReindex()
-	content, md, err := v.magnify(hash, MagnifyOpts{
+	content, md, err := v.magnify(hash, magnifyOpts{
 		DisplayName: req.GetDisplayName(),
 		Tags:        req.GetTags(),
 		Reindex:     reindex,
@@ -227,123 +220,4 @@ func (v *V2) Remove(ctx context.Context, req *lensv2.RemoveReq) (*lensv2.RemoveR
 	}
 
 	return &lensv2.RemoveResp{}, nil
-}
-
-// MagnifyOpts declares configuration for magnification
-type MagnifyOpts struct {
-	DisplayName string
-	Reindex     bool
-	Tags        []string
-}
-
-func (v *V2) magnify(hash string, opts MagnifyOpts) (content string, metadata *models.MetaDataV2, err error) {
-	if v.se.IsIndexed(hash) && !opts.Reindex {
-		return "", nil, fmt.Errorf("object '%s' has already been indexed", hash)
-	}
-
-	// set up args
-	if opts.Tags == nil {
-		opts.Tags = make([]string, 0)
-	}
-
-	// set up logger
-	var l = logs.NewProcessLogger(v.l, "magnify", "hash", hash)
-	var start = time.Now()
-	defer func() { l.Infow("magnification ended", "duration", time.Since(start)) }()
-
-	// retrieve object and detect content type
-	contents, err := v.px.ExtractContents(hash)
-	if err != nil {
-		return "", nil, fmt.Errorf("failed to find content for hash '%s'", hash)
-	}
-	contentType := http.DetectContentType(contents)
-	if contentType == "" {
-		return "", nil, fmt.Errorf("unknown content type for document '%s'", hash)
-	}
-	l.Infow("object retrieved and content type detected",
-		"content_type", contentType)
-
-	// contentType will be in the format of `<content-type>; charset=...`
-	// we use strings.FieldsFunc to seperate the string, and to be able to exmaine
-	// the content type
-	var parsed = strings.FieldsFunc(contentType, func(r rune) bool { return (r == ';') })
-	if parsed == nil || len(parsed) == 0 {
-		return "", nil, fmt.Errorf("invalid content type '%s'", contentType)
-	}
-
-	// scrape for content based on content-type
-	var category MimeType
-	switch parsed[0] {
-	case "application/pdf":
-		category = MimeTypePDF
-		text, err := v.oc.Analyze(hash, contents, "pdf")
-		if err != nil {
-			return "", nil, err
-		}
-		content = text
-	default:
-		var parsed2 = strings.FieldsFunc(contentType, func(r rune) bool { return (r == '/') })
-		if parsed2 == nil || len(parsed2) == 0 {
-			return "", nil, fmt.Errorf("invalid content type '%s'", contentType)
-		}
-		switch parsed2[0] {
-		case "text":
-			category = MimeTypeDocument
-			content = string(contents)
-		case "image":
-			category = MimeTypeImage
-			keyword, err := v.tf.Analyze(hash, contents)
-			if err != nil {
-				l.Warnw("failed to categorize image", "error", err)
-				return "", nil, errors.New("failed to categorize image")
-			}
-
-			// grab any text in image
-			text, err := v.oc.Analyze(hash, contents, "image")
-			if err != nil {
-				l.Warnw("failed to OCR image", "error", err)
-				content = keyword
-			} else {
-				content = text
-			}
-			opts.Tags = append(opts.Tags, keyword)
-		default:
-			return "", nil, errors.New("unsupported content type for indexing")
-		}
-	}
-
-	return content, &models.MetaDataV2{
-		DisplayName: opts.DisplayName,
-		MimeType:    contentType,
-		Category:    string(category),
-		Tags:        opts.Tags,
-	}, nil
-}
-
-// Store is used to store our collected meta data in a formatted object
-func (v *V2) store(hash string, content string, md *models.MetaDataV2) error {
-	return v.se.Index(engine.Document{
-		Object: &models.ObjectV2{
-			Hash: hash,
-			MD:   *md,
-		},
-		Content: content,
-	})
-}
-
-// Update is used to update an indexed object
-func (v *V2) update(hash string, content string, md *models.MetaDataV2) error {
-	if err := v.remove(hash); err != nil {
-		return err
-	}
-	return v.store(hash, content, md)
-}
-
-// Remove is used to remove an indexed object
-func (v *V2) remove(hash string) error {
-	if !v.se.IsIndexed(hash) {
-		return fmt.Errorf("object '%s' does not exist", hash)
-	}
-	v.se.Remove(hash)
-	return nil
 }
